requirements-froze.txt
agora-token-builder==1.0.0
apturl==0.5.2
arabic-reshaper==2.1.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.2.10
asn1crypto==1.5.1
asttokens==2.0.8
async-generator==1.10
attrs==21.4.0
autopep8==2.0.0
Babel==2.8.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
bleach==5.0.1
blinker==1.4
Brlapi==0.8.3
Brotli==1.0.9
certifi==2020.6.20
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
cliapp==1.20180812.1
click==8.1.3
cmdtest==0.32+git
colorama==0.4.4
coloredlogs==15.0.1
command-not-found==0.3
crayons==0.4.0
cryptography==38.0.1
cssselect2==0.7.0
cupshelpers==1.0
customtkinter==4.6.3
darkdetect==0.7.1
dbus-python==1.2.18
debugpy==1.6.3
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
diffusers==0.3.0
distlib==0.3.4
distro==1.7.0
distro-info===1.1build1
Django==3.0.8
django-admin-interface==0.24.1
django-colorfield==0.8.0
django-cors-headers==3.13.0
django-crispy-forms==1.9.1
django-rest-knox==4.2.0
django-rest-passwordreset==1.3.0
djangorestframework==3.13.1
duplicity==0.8.21
entrypoints==0.4
executing==0.10.0
fasteners==0.14.1
fastjsonschema==2.16.1
filelock==3.4.2
Flask==2.0.1
flatbuffers==2.0.7
future==0.18.2
google-trans-new==1.1.9
googletrans==4.0.0rc1
gTTS==2.2.3
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.12.1
html5lib==1.1
httpcore==0.9.1
httplib2==0.20.2
httpx==0.13.3
huggingface-hub==0.9.1
humanfriendly==10.0
hyperframe==5.2.0
idna==2.10
imageio==2.31.0
imageio-ffmpeg==0.4.8
importlib-metadata==4.6.4
invisible-watermark==0.1.5
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==8.0.1
itsdangerous==2.1.0
jedi==0.18.1
jeepney==0.7.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.14.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter_client==7.3.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.2
kaitaistruct==0.9
kazam==1.4.5
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
ldap3==2.8.1
lockfile==0.12.2
louis==3.20.0
lxml==4.9.2
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib-inline==0.1.6
mistune==2.0.4
mitmproxy==6.0.2
monotonic==1.6
more-itertools==8.10.0
moviepy==1.0.3
mpmath==1.2.1
msgpack==1.0.3
nbclient==0.6.7
nbconvert==7.0.0
nbformat==5.4.0
neattext==0.1.3
nest-asyncio==1.5.5
netifaces==0.11.0
notebook==6.4.12
numpy==1.22.1
oauthlib==3.1.0
olefile==0.46
onnx==1.12.0
onnxruntime==1.12.1
opencv-python==4.7.0.72
oscrypto==1.3.0
outcome==1.1.0
pandas==1.4.2
pandocfilters==1.5.0
paramiko==2.9.3
parso==0.8.3
passlib==1.7.4
pexpect==4.8.0
pickleshare==0.7.5
piexif==1.1.3
Pillow==9.2.0
pipenv==2022.1.8
platformdirs==2.4.1
powerline-status==2.7
proglog==0.1.10
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.12.4
psutil==5.9.1
ptyprocess==0.7.0
publicsuffix2==2.20191221
pure-eval==0.2.2
pyasn1==0.4.8
pycairo==1.20.1
pycodestyle==2.10.0
pycparser==2.20
pycups==2.0.1
pydub==0.25.1
Pygments==2.13.0
PyGObject==3.42.1
pyHanko==0.15.1
pyhanko-certvalidator==0.19.7
pyinotify==0.9.6
PyJWT==2.6.0
pymacaroons==0.13.0
PyMySQL==1.0.2
PyNaCl==1.5.0
pyngrok==5.1.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
PyPDF2==1.26.0
PyPDF3==1.0.6
pyperclip==1.8.2
pyRFC3339==1.1
pyrsistent==0.18.1
pytesseract==0.3.8
python-apt==2.4.0+ubuntu1
python-bidi==0.4.2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==0.19.0
python3-openid==3.2.0
pytz==2020.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
pyxdg==0.27
PyYAML==6.0
pyzmq==23.2.1
qrcode==7.3.1
qtconsole==5.3.1
QtPy==2.2.0
regex==2022.9.13
reportlab==3.6.12
requests==2.24.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rfc3986==1.5.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
schedule==1.1.0
scikit-learn==1.1.2
scipy==1.9.1
SecretStorage==3.3.1
selenium==4.1.0
Send2Trash==1.8.0
simplejson==3.17.6
six==1.15.0
sklearn==0.0
sniffio==1.2.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sortedcontainers==2.4.0
soupsieve==2.3.1
SpeechRecognition==3.8.1
sqlparse==0.3.1
ssh-import-id==5.11
stack-data==0.5.0
stegano==0.11.2
svglib==1.4.1
sympy==1.11.1
systemd-python==234
tabula-py==2.3.0
terminado==0.15.0
tesseract==0.1.3
threadpoolctl==3.1.0
tinycss2==1.2.1
tokenizers==0.12.1
tomli==2.0.1
torch==1.12.1
tornado==6.2
tqdm==4.64.1
traitlets==5.3.0
transformers==4.19.2
trio==0.19.0
trio-websocket==0.9.2
ttystatus==0.38
tweepy==4.4.0
typing_extensions==4.3.0
tzdata==2022.7
tzlocal==4.2
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uritools==4.0.0
urllib3==1.25.9
urwid==2.1.2
usb-creator==0.3.7
virtualenv==20.13.0
virtualenv-clone==0.5.7
wadllib==1.3.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
widgetsnbextension==4.0.2
wifite==2.2.5
wsproto==1.0.0
xdg==5
xhtml2pdf==0.2.8
xkit==0.0.0
zipp==1.0.0
